requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anaconda==0.0.1.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
argcomplete==1.10.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
autopep8==2.0.4
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.8.2
bidict==0.22.1
bleach==6.1.0
blinker==1.6.3
cachelib==0.12.0
cachetools==5.3.1
certifi @ file:///private/tmp/python-certifi-20240201-4765-6ufe0c/certifi-2024.2.2
cffi @ file:///private/tmp/cffi-20231014-5567-1hyue8f/cffi-1.16.0
chainlit==0.7.301
chardet==3.0.4
charset-normalizer==3.3.1
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
compressed-rtf==1.0.6
contourpy==1.1.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
Django==5.0.2
djangorestframework==3.15.1
docx2txt==0.8
ebcdic==1.1.1
executing==2.0.0
extract-msg==0.28.7
faiss-cpu==1.7.4
fastapi==0.100.1
fastapi-socketio==0.0.10
fastjsonschema==2.18.1
filelock==3.12.4
filetype==1.2.0
Flask==2.2.5
Flask-Session==0.4.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
glob2==0.7
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
i==2.1.0
identity==0.2.0
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
InstructorEmbedding==1.0.1
ipykernel==6.26.0
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.14.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.150
langsmith==0.0.49
Lazify==0.4.0
libclang==16.0.6
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
msal==1.27.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.1
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.16.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
opencv-python==4.8.1.78
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opt-einsum==3.3.0
overrides==7.4.0
packaging @ file:///private/tmp/python-packaging-20231018-5368-zv724j/packaging-23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathspec @ file:///private/tmp/python-pathspec-20231212-4946-k7fih8/pathspec-0.12.1
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.10.24
platformdirs==3.11.0
posthog==3.0.2
prisma==0.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil @ file:///private/tmp/python-psutil-20240119-4673-f71z61/psutil-5.9.8
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser @ file:///private/tmp/pycparser-20231013-5787-1anpge2/pycparser-2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments @ file:///private/tmp/pygments-20231121-5274-1rlyumb/pygments-2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==3.16.4
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.8.0
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-multipart==0.0.6
python-pptx==0.6.22
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML @ file:///private/tmp/pyyaml-20231018-5610-1rdh755/PyYAML-6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
simple-websocket==1.0.0
six==1.12.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.2
style==1.1.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
terminado==0.17.1
textract==1.6.5
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.34.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
update==0.0.1
uptrace==1.20.2
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
virtualenv==20.24.6
watchfiles==0.20.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.1.9
yarl==1.9.2
zipp==3.17.0
